Restructure and remove unused imports
